Updated to editor type app
