--- app/.idea/.name
@@ -1,1 +1,0 @@
-__init__.py
