+++ viktor.config.toml
@@ -1,1 +1,1 @@
-app_type ="tree"
+app_type ="editor"
